api/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
Django==5.1.4
djangorestframework==3.15.2
mysqlclient==2.2.6
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
sqlparse==0.5.3
typing_extensions==4.12.2
